+++ Sources/PrintingKit/PrintItem.swift
@@ -32,26 +32,13 @@ public enum PrintItem {
     
     /// A PDF document file at a certain URL.
     case pdfFile(at: URL?)
+    
+    /// A plain string with a PDF page configuration.
+    case string(String, configuration: Pdf.PageConfiguration = .standard)
 }
 
 @available(iOS 16.0, macOS 13.0, *)
 public extension PrintItem {
-    
-    /**
-     Build an ``PrintItem/attributedString(_:configuration:)``
-     print item with a plain string.
-     
-     - Parameters:
-       - string: The string to print.
-       - configuration: The page configuration to use, by default `.standard`.
-     */
-    static func string(
-        _ string: String,
-        configuration: Pdf.PageConfiguration = .standard
-    ) -> PrintItem {
-        let str = NSAttributedString(string: string)
-        return .attributedString(str, configuration: configuration)
-    }
     
     /**
      Try to create an ``PrintItem/imageData(_:)`` print item